Add UI integration to Lacros tests.
This CL adds a precondition for Lacros with UI integration.
It will continue to use the binary downloaded from Google Storage.
To do this, the permissions of the binary path needed to be loosened.

When launching from UI I have tried to keep the default launching as
close to default as possible and only added args that enables the test
extensions.

A followup CL (next in the chain) will launch lacros from the UI.

Also extracted the chrome args passed to preconditions to a var
instead of duplicating.

BUG=chromium:1149232
TEST=lacros.Basic still works

Change-Id: I87be63e5a808434fdd56a5694e46d35fdcdcaa5d