Add a lacros version of launcher.SearchBuiltInApps.
This test searches for an app in the launcher and checks that it opens.

I needed to share the lacros library with the sysui-launcher package. Since it
was defined in cros/ui I was not able to use it directly. Instead I have
merged the two lacros.go files into one that all tests can use. I think
it makes the code clearer overall too.

BUG=chromium:1149233
TEST=launcher.SearchBuiltInApps.lacros, launcher.SearchBuiltInApps
TEST=ui.LauncherAnimationPerf.lacros, ui.OverviewPerf.lacros,
ui.WindowCyclePerf.lacros

Change-Id: If6f75411b59d97d4764eb9549a598dbb9179a723
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/2552938
Tested-by: David Haddock <dhaddock@chromium.org>
Commit-Queue: David Haddock <dhaddock@chromium.org>
Reviewed-by: Brendan Hansknecht <bhansknecht@chromium.org>