video: Add platform encoding tests
This CL adds a wrapper for vp8enc, the libva-utils binary
to run VP8 encoding, in 3 variants (QVGA, VGA and HD).
It uses a similar approach to video.EncodeAccel* in that
it decodes a webm file to be used as input (otherwise the
files would be too large). This CL also parses the output
and produces the resulting FPS as a Perf metric.

BUG=b:174103282
TEST=video.PlatformEncode.* on Celes

Change-Id: I554fcd3f7687a2e3076d713fd3d2b00081bde63a