Add RV22 Tast-test for WM ARC++ R
Implement RV22: resizable/conversion: split screen.
In RV22 two activities are started, the first one is referred to as
underActivity and the second one as overActivity. A swipe from the
top middle of the screen to the center is simulated, this will take
the overActivity to a snapping state. Then it will be snapped to the
left. UnderActivity will be snapped to the right by a touch on the
center of the right side of the screen.Then window bounds for each
activity will be asserted. Then tablet mode will be disabled and
window bounds will be asserted in clamshell mode. The display will be
rotated to portrait mode. Left snapped activity (overActivity) will
be snapped to the top and the other one to the bottom and window
bounds will be asserted. By disabling the tablet mode, the device
will convert to clamshell mode. The overActivity must be snapped to
the left and the underActivity to the right. Then tablet mode will be
enabled again. At the end the window bounds will be asserted at
portrait tablet mode.
For more info please visit go/arc-wm-r.

BUG=b:140203139
TEST=tast run $DUT arc.WMResizableConversion

Change-Id: I233e44b133ccdbda3cb9bf4d57dade28f8893ee2
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/2568815
Tested-by: Armen Keshishian <armenk@google.com>
Reviewed-by: Po-Hsien Wang <pwang@chromium.org>
Commit-Queue: Armen Keshishian <armenk@google.com>