tast-tests: Clean battery metrics logic to expect only one battery
Before this CL, battery metrics logic was implemented to be able to
handle multiple batteries, but it is clear now that we have at most one
battery for each device. To simplify the logic, this CL changes some
interfaces to take a single string of a path to battery sysfs.

BUG=None
TEST=tast run $DUT arc.PowerIdlePerf.vm

Change-Id: I5ca52bb831dfc075699f49be8dddcedc5149d0b3