tast-tests: Clarify "device has multiple batteries" error
ReadBatteryStatus/Capacity checks the number of battery and expects
there is only one battery. If not, it raised "device has multiple
batteries" error but the message was confusing and not suitable for
the case which there is no battery. This CL makes the message more
clearer and adds the actual number of batteries to a error.

BUG=None
TEST=tast run $DUT arc.PowerIdlePerf.vm

Change-Id: I5ca52bb831dfc075699f49be8dddcedc5149d0b3