wifi: BeaconInterval: Sync the channel setting with Autotest.
When ported to Tast, the test uses CommonAPOptions instead of the
setting in Autotest. However, on some devices, 5GHz band has no IR
setting and cannot send probe requests. With a longer beacon interval,
DUT is more prone to miss beacons in passive scans and making the test
fail with "cannot find matching service".

Use channel 6 as the Autotest version to deflake the test.

BUG=b:173633813
TEST=`tast run $DUT wifi.BeaconInterval` a few times on a trogdor DUT

Change-Id: I9c06f33ec75c66ce4d9b2962bc8a530fdc42a4ad