new test: ArcGamePerformanceTest
This is a direct copy&paste from:
http://cs/pi-arc-dev/frameworks/base/tests/GamePerformance/

Even though this test lives in frameworks/base, this test was created for ARC.
So moving it to ChromeOS Tast repo makes sense since it is being used by
Tast test.

Patchset#0: Contains the original, unmodified files
Patchset#1-3: Fixes to make it compile with AOSP SDK 27.
Patchset#4: Reformatted to comply with Android Java guidelines

BUG=b:174712817
TEST=emerge-$BOARD tast-local-apks-cros
     Install APK. Verified that it works Ok.

Change-Id: Ic4d42ec5f4616b481afbae77ee66bbac50470d85