firmware.BootMode: Reset power-state at test end
Currently, if the firmware.BootMode test ends with the DUT powered-off,
then all future tests in the same suite will fail to connect, and the
Tast harness will fail to write results. See attached bug for
details/examples.

This CL reserves some time at test-end to reset the power-state in such
scenarios, so that future tests will be unaffected.

BUG=b:174594888
TEST=firmware.BootMode.normal
TEST=firmware.BootMode.normal, with RebootToMode modified to fail

Change-Id: I9ba95ec70a4cd6d2a74a47af9ed9960b4ccaa96a