tast-tests: Add chrome precondition that disables sync
shelf.LaunchedApps verifies that the default pinned apps can be launched
via the Shelf. When the chrome.LoggedIn precondition finishes executing,
the default apps are not being pinned leading to only the Chrome app
being shown in the shelf.

This only happens intermittently and contributes to the flakiness. Given
the account used by the chrome.LoggedIn precondition is actually not a
real user flow, disabling syncing stops this race condition and always
pins the default apps.

BUG=chromium:1154486
TEST=./tools/repeat-test.sh -s -r 100 -- <vayne-dut> shelf.LaunchedApps

Change-Id: Iec3bb2ae26dba2a94b965b00a794ee77fc66cfce