tast-tests: Disable sync when using fake login on Chrome
The fakeLogin flow for Chrome logs in with testuser@gmail.com and
bypasses authentication, this leads to the user profile sync system not
working properly. For any chrome login flow that uses fakeLogin, launch
with the --disable-sync flag added.

shelf.LaunchedApps checks the default pinned apps can be launched via
the shelf. When the sync system does not report properly, the default
apps are not being pinned and the fallback is just having Chrome showing
up.

BUG=chromium:1154486
TEST=./tools/repeat-test.sh -s -r 100 -- <vayne-dut> shelf.LaunchedApps

Change-Id: Iec3bb2ae26dba2a94b965b00a794ee77fc66cfce