tast-tests: Disable sync in shelf.LaunchedApps on Chrome
The fakeLogin flow for Chrome logs in with testuser@gmail.com and
bypasses authentication, this leads to the user profile sync system not
working properly. Expose a new precondition LoggedInDisableSync which
simply passes the --disable-sync flag.

shelf.LaunchedApps checks the default pinned apps can be launched via
the shelf. When the sync system does not report properly, the default
apps are not being pinned and occasionally the only pinned app is
Chrome.

BUG=chromium:1154486
TEST=./tools/repeat-test.sh -s -r 100 -- <vayne-dut> shelf.LaunchedApps

Change-Id: Iec3bb2ae26dba2a94b965b00a794ee77fc66cfce