wifi.ChannelScanDwellTime: Refine error of beacon not found in pcap.
The test failed sometimes with reason like:
"Dwell time 2562047h47m16.854775807s is not within range [5ms, 250ms]"

The large dwell time is from subtracting zero value of the time.Time in
map and the cause is that we cannot find timestamp of the beacon in the
pcap file.

Refine the error string so the problem is better stated.

BUG=b:171600107
TEST=tast run $DUT wifi.ChannelScanDwellTime

Change-Id: Ia6ef8713e1825aee63b613f977e2038083ef970f