platform.BootPerf: make WiFi metric Recommended
We expect all Chrome OS hardware to have WiFi, but there are occasions
(especially bad lab DUTs) where the WiFi hardware is flaky. If we were
to mark WiFi-related events as Required, then failures would obstruct
reporting other metrics which were successfully collected. If we leave
this metric as optional (Required==false), then we'll only report WiFi
metrics if we manage to collect a stat before all the other Required
events are complete.

Neither situation is ideal, so we generate a 3rd option --  Recommended
-- which can be used to perform a best-effort collection.

Alternatives could include adding separate test cases just to track
WiFi, but given the length of this test (several minutes), that option
is probably too expensive.

BUG=b:172212439
TEST=platform.BootPerf, with "flaky" (artificially disabled) WiFi

Change-Id: Ie62530f35857137118320d969dd4deb699d5057a