Add support for enabling Lacros by component updater (omaha).
Currently all lacros tests will download a binary from Google Storage before the
test. To test the flow that end users use, we will add a precondition that
downloads the lacros binary from omaha.

Restrict the test to only run on a couple of boards so as not to
overwhelm the lab with lots of downloads.

BUG=chromium:1149232
TEST=lacros.Basic, lacros.ShelfLaunch, lacros.ShelfLaunch.omaha

Change-Id: I447d67a34785b65af4b98694ba78014e082bd47c
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/2536739
Tested-by: David Haddock <dhaddock@chromium.org>
Reviewed-by: Brendan Hansknecht <bhansknecht@chromium.org>
Commit-Queue: David Haddock <dhaddock@chromium.org>