Fix the waiting logic of ui.TabletOperations
Flakiness is seen on the subscenario of 'overview-splitview' in
ui.TabletOperations on certain devices. As I checked the order
of manipulations of overview mode, I've come to realize that
sometimes the test scenario does not wait properly for the
ending overview animation in the splitview.

Currently, the test scenario generates some gesture events
and then wait for the end of the window animation, assuming
that the window animation is ongoing. Sometimes the window
animation is not yet started, and therefore the waiting finishes
immediately, and then next the test scenario invokes
ash.SetOverviewModeAndWait which cancels the overview animation.

Rather, the test scenario should check the overview mode to be
finished first. That way it properly ensures that the side
overview finished.

BUG=chromium:1142067
TEST=run ui.TabletOperations on volteer / delbin for 6 times

Change-Id: Ided6b3a60a6ec38d305f1c4d0a46fc21868826b9
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/2573380
Tested-by: Jun Mukai <mukai@chromium.org>
Reviewed-by: Xiyuan Xia <xiyuan@chromium.org>
Commit-Queue: Jun Mukai <mukai@chromium.org>