wifi: SSIDSwitchBack: Fix a bug in context usage.
The test sets 30s timeout for waiting for reconnection. However, the
timeout should not limiting the following verification and teardown.
This makes the test flaky when reconnection takes most of the 30s
timeout and there might be no enough time for VerifyConnection.

Properly set the context so that only the timeout only applies in the
waiting step.

BUG=b:173175375
TEST=tast run $DUT wifi.SSIDSwitchBack

Change-Id: I63164d85d69940c83a42464adae19b05f38fd441
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/2577140
Tested-by: Yen-lin Lai <yenlinlai@google.com>
Reviewed-by: Shuo-Peng Liao <deanliao@chromium.org>
Commit-Queue: Yen-lin Lai <yenlinlai@google.com>