tast-tests: Disable arcvm-console.txt on ARM devices
The commit at https://crrev.com/c/2232379/ added ARCVM kernel's console
output to tast test results as a log file `arcvm-console.txt`. This
feature requires that pstore is enabled on the Android kernel of ARCVM,
and the pstore feature is enabled only on x86 devices, so the commit
breaks tests on ARM devices. This commit disables `arcvm-console.txt` on
ARM devices and fixes the issue.

BUG=b:149921859
TEST=run `tast run $DUT arc.Boot.vm` on an ARM device

Change-Id: Iea5fb1ba05fe66d1e0bba4c5ba19ccd61bd554d0