tast-tests: Add "arc_pstore" to SoftwareDeps of arc.KernelCrash.*
The guest kernel crash reporting of ARCVM depends on the pstore feature
of Linux kernel. Now pstore is not enabled on ARM devices and tests are
failing on them. So, this commit disables the test on ARM devices.

Also this commit replace "amd64" feature of arc.VmPstoreDump with "arc_pstore".

BUG=b:139650101, b:175089029
TEST=run `tast run my_crbook arc.KernelCrash.{real,mock}_consent` on a x86_64 device and an ARM device.

Cq-Depend: chromium:2578619
Change-Id: Iff8a50153c6169704efd13a1b142bbd6a8d12fc8