Fix wifi.PersistenceBluetoothSansWifi init flake
tast.wifi.PersistenceWifiSansBluetooth is currently falling below 90%
pass rate. One issue is that the test fails if bluetooth is not enabled
when the test starts. This shouldn't cause the test to fail.

Another issue comes from the fact the test fails when it checks if the adapter
is down. We now poll on the assert check after the adapter is disabled.

BUG=b:175047269
TEST=tast run ${DUT} wifi.PersistenceWifiSansBluetooth

Change-Id: Ia008e7fd4c54534fc97cd6d8049f35db832fde07
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/2577851
Tested-by: Billy Zhao <billyzhao@chromium.org>
Reviewed-by: Shuo-Peng Liao <deanliao@chromium.org>
Reviewed-by: Yen-lin Lai <yenlinlai@google.com>
Commit-Queue: Billy Zhao <billyzhao@chromium.org>