Speculative fixes of ui.TabletOperations flakiness
There are still a few rare failures on:
- "failed to find the Chrome icon"
- Ash.HotseatWidgetAnimation.Widget.AnimationSmoothness.
  TransitionToHiddenHotseat is missing

Speculatively, the former could be because Chrome icon may not
be visible to automation API; FindWithTimeout should fix the
problem.

For the latter, a theory is that the Chrome icon is still animating
(it is in a hotseat, which is revealed by the gesture). In that
case, the touch will fall into a wrong location, and thus does
not cause any changes, but WaitWindowFinishAnimating is succeeding
as no window is animating. For this, WaitLocationStable will
stabilize this. Also, check the window status to make sure that
the window is actually resumed.

BUG=chromium:1156298
TEST=tast run ui.TabletOperations 10 times

Change-Id: Ie3073768ca854c820e0f7388171575814f159200