A speculative fix of ui.DesktopControl flakiness
It fails during ui.Find for ash/StatusAreaWidgetDelegate. It is
not quite clear how it happens. Possibly, some of the previous
tests may change the shelf visibility which also affects the
status area widget visibility, and so the automation API may
cache an older state when the status area widget is not visible.

Because of that:
- make sure that the shelf behavior to be 'always', as it is
  the expected situation of this test
- use ui.FindWithTimeout and StableLeftClick in case that the
  automation API keeps a stale status of the shelf visibility.

BUG=chromium:1156333
TEST=tast run ui.DesktopControl 10 times

Change-Id: I0d25267fd7db98e52ea039ec90b402281fd8d2b5