FWTries: Reset power if disconnected at test end
Sometimes the reboots cause an error, leaving the DUT powered-off at
test end. This prevents future tests from running, and prevents the test
harness from reporting test results to Tauto. In order to address that
issue, this CL defers a power-reset via Servo if the DUT is disconnected
at end-of-test.

BUG=b:162367972
TEST=firmware.FWTries
TEST=firmware.FWTries, modified to disconnect from DUT before test-end

Change-Id: I1e019fea807bb7c115c5df2d467b7b92ca995f9d