security.SandboxedServices: add cryptohome-namespace-mounter baseline.
'cryptohome-namespace-mounter' is the process that actually performs
the mounts that make user data available. The test restarts the UI job
to avoid this kind of flake, but clearly it's not enough. Let's add a
baseline entry for this process.

Per the failing test run logs,
cryptohome-name uid=0 gid=0 pidns=4026531836 mntns=-1 nnp=false seccomp=false ecaps=0x3fffffffff

(Logs at
https://ci.chromium.org/ui/p/chromeos/builders/vmtest/amd64-generic-direct-tast-vm/b8861890299483247616/overview)

So expecting it to run as root:root is accurate (and matches the code.)

BUG=b:174865908
TEST=Passes on betty-pi-arc (but this doesn't mean much.)

Change-Id: Ia530f4706ef525889dfec6805bfd788e3e8354ea