firmware: Add test to exercise bio_fw_updater
This test flashes an out-dated firmware binary, then call bio_fw_updater
and verify the update succeeds. An alternative way is to trigger
bio_fw_updater with a reboot.

This change also moves the logic for reading updater logs into a service
so that it can be shared with the local test
firmware.FpUpdaterSucceeded.

BUG=b:169958710
TEST=tast -verbose run -ephemeraldevserver=false <IP> firmware.FpUpdater
TEST=tast -verbose run -ephemeraldevserver=false <IP>
	firmware.FpUpdaterSucceeded
TEST=tast -verbose run -var servo=localhost:9999
	-ephemeraldevserver=false <IP> firmware.FpSensor

Change-Id: Id3531b8987aaa7f7fea000f0a27178420981f6ca
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/2595471
Tested-by: Yicheng Li <yichengli@chromium.org>
Reviewed-by: Tom Hughes <tomhughes@chromium.org>
Reviewed-by: Shuhei Takahashi <nya@chromium.org>
Commit-Queue: Yicheng Li <yichengli@chromium.org>