A speculative fix of ui.WindowControl failure.
Occasionally 'resize' subscenario fails due to the missing
histogram of Ash.InteractiveWindowResize.TimeToPresent. That
means the subscenario runs fine but does not generate the
performance data.

That suggests that the location of mouse events were somehow
incorrect and did not cause any window resizes.

Then hypothetically ash.FindWindow may be not great as the
window may be in some transitive status. This CL replaces it
by ash.WaitForCondition and makes sure that the window is in
a normal state and can be resized.

BUG=chromium:1156762
TEST=ui.WindowControl

Change-Id: I22bf2325e434c98378f343cd95f65a3f16557782
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/2579669
Reviewed-by: Xiyuan Xia <xiyuan@chromium.org>
Commit-Queue: Jun Mukai <mukai@chromium.org>
Tested-by: Jun Mukai <mukai@chromium.org>