ui.QuickSettingsLockScreen: unlock screen at the end of the test
Previously the test would leave the DUT with the screen locked. It
doesn't use the chrome.LoggedIn precondition so I thought the state
would be cleared for subsequent tests, but this turned out to not be the
case. crbug/1156812 describes the issue we encountered while trying to
land CL:2559539. Renaming the quick settings tests changed their running
order, and the DUT being in the locked state caused a subsequent test
(security.SELinuxProcesses) to fail. For the time being, this test
should take care to clean up the state as best as it can until
crbug/1156812 is resolved.

BUG=chromium:1156914
TEST=tast run localhost:2222 ui.QuickSettingsLockScreen

Change-Id: I44fead707271c0074b1ea35873ae92cf88d98b94