tast-tests/video: per-codec args and regexp for PlatformEncoding
PlatformEncoding constructs the arguments for the command line
binary in the main function. This CL splits it into a function
so that different codecs can have different command lines. It
also makes the regexp used for FPS parsing configurable.

No functionality changes intended with this CL (see the next one
for it).

BUG=b:174103282
TEST=video.PlatformEncoding.* on Celes

Change-Id: Ic5d1fe83c528dafe043b855ce09701c4b102e13e