tast-tests: Migrate external ARC app to internal for arc.Sharesheet
arc.Sharesheet uses an external Photo Editor app which required setting
up the Play Store and was a source of flakiness. ARC team provides a way
to bundle minimal test apps that are sideloaded into ARC to verify
functionality, see:

  //android/README.md

for more information. The minimal app in this CL checks that the file
that was passed in is readable by ARC apps and verify the end-to-end
flow of data makes it way through the Sharesheet. Given we are
sideloading the app, we don't need the Play Store dependency anymore.

BUG=chromium:1156310, chromium:1153218
TEST=emerge tast-local-apks-cros && cros deploy
TEST=tast run <DUT-IP> arc.Sharesheet

Change-Id: I67884b570dafebe693477beeea0ce8ea980c3f52