arc.TouchPerf: add touch screen move test by replay kernel events
Touch.go shows higher latency for move events, so simulate touch screen
move events by evemu-play.

Since it's not able to calculate the time from host to ARCVM for each of
the events when using replay, uses the mean time diff instead. Since the
average diff is ~1 ms, it can be used roughly when considering the move
latency is above 6 ms.

BUG=b:173453054
TEST=tast -verbose run <DUT> arc.TouchPerf(.vm) pass

Change-Id: Idbb13547ec5f6b010ec03f72e1fc9b57c26ea7ad