Fix wifi.BluetoothXorWifi init flake
tast.wifi.BluetoothXorWifi is currently more flaky than its autotest
counterpart. The majority of the flake comes from failing on the
assumption that Bluetooth must be enabled when the test starts. This
patch changes the behavior such that if Bluetooth adapter is disabled, it just
proceeds to just disable all the phys instead of failing immediately.

BUG=b:175538931
TEST=tast run ${DUT} wifi.BluetoothXorWifi

Change-Id: I4a1bcb835e32b01a8def26ab331bd7a642cf37fc