Fix ui.WindowControl flakiness
I checked the behavior of problematic builds (trembyle), and
noticed that sometimes the resizing window can be overlapping
with the shelf, which prevents the resizing behavior indeed.

Make the resizing from the top-right instead of the bottom right,
as top-right would be a safer place for window resizing.

BUG=chromium:1156762
TEST=ui.WindowControl on trembyle

Change-Id: Ie12bf9ff64b7358778d3e7cfd60def1d14015aa3
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/2594254
Tested-by: Jun Mukai <mukai@chromium.org>
Reviewed-by: Xiyuan Xia <xiyuan@chromium.org>
Commit-Queue: Jun Mukai <mukai@chromium.org>