Reform CPU utilization collector
Add the record of frequency / max frequency. With that, per-CPU
data would also be necessary.  Revisiting the per-CPU utilization,
the existing gopsutil/cpu utilities may not fit well.
cpu.PercentWithContext has the parameter of "percpu" flag, though
it does not care about which core is online/offline at this time.
When some cores turn online additionally, this will confuse the
timeline data. Thus now cpu.go computes the duration by itself.

As this is very specific to the CUJ tests, the collector is
moved back to cuj directory. Along with that, moves mem.go back
to cuj too and integrates into the memory-diff collector.

BUG=chromium:1158047
TEST=run CUJ tests

Change-Id: I786b27f4b9453bc6b86cc010e38b7721048d5c97