tast-tests: Add ARCVM console output to test results with fixtures again
This is a reland of 692a6980af631a4277af9a6c9ec43c136de83354 with an
additional check of the output of vm_pstore_dump command.

In the PostTest hook of "arcBooted" fixture, the console output of ARCVM
guest kernel is collected as a log file. This use vm_pstore_dump command
to read /run/arcvm/arcvm.pstore file, but this command fails when the
file doesn't exist. We need the additional check.

BUG=b:177460986
TEST=Run tests and check /tmp/tast/results/latest/tests/*/arcvm-console.txt after the tests finished.
TEST=Run two or more tests at once. e.g. $ tast run my_crbook arc.BuildProperties.vm arc.CheckAndroidVersion.vm
TEST=Run tests and remove /home/root/*/crosvm/*.pstore during the tests running.

Change-Id: I0a1bb67dfd378e0b9e8522a475a5b1b85ad7d058