tast-tests: Swipe faster in DesksTrackpadSwipePerf
Seemed to be failing on candy device because the swipe
did not pass threshold to trigger desk change. Threshold
is in "touchpad units" which depends on both swipe
distance and speed. This CL speeds up the swipe a bit to
pass for candy, but may need further tweaking.

BUG=chromium:1149643
TEST=tast -verbose run <IP> ui.DesksTrackpadSwipePerf

Change-Id: Ic9fe878b9edfac6d519fffd568fe8d78f145e8b3
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/2595476
Tested-by: Sammie Quon <sammiequon@chromium.org>
Reviewed-by: Jun Mukai <mukai@chromium.org>
Commit-Queue: Sammie Quon <sammiequon@chromium.org>