tast-test: Add namespace check precondition to cryptohome.GuestSession
The code changes in chromium:2586593 and chromium:2586594 move the mount
namespace creation from cryptohome to session_manager. There is no
direct check mechanism in cryptohomed for the mount namespace so the
test may fail due to a race condition although it is a very small
probability.

We add a precondition check in the cryptohome.GuestSession test. This
precondition check only prints out an error message if the mount
namespace is not created yet. This way we will know the reason of the
test failure.

Also note that this error message will be printed for successful tests
as well for a short time period until the CLs mentioned above will be
landed.

BUG=chromium:1132444
TEST=tast run DUT cryptohome.GuestSession prints
the following messages (without chromium:2586593 and chromium:2586594):
'user session mount namespace has not been created at
/run/namespaces/mnt_chrome'
'Mount namespace is not ready.'

Change-Id: I339609c939efd09b29bb3b02c4508ff7b0d4b8f6