tast-test: Add namespace check precondition to cryptohome.GuestSession
The code changes crrev/2586593 and crrev/2586594 moves the namespace
creation from cryptohome to session_manager. With those changes there
will be no direct control mechanism in cryptohome for the mount
namespace so the test may fail due to a race condition (although it
is a very small possibility).
We add a precondition check in the cryptohome.GuestSession test. This
precondition check only prints out an error message if the mount
namespace is not created yet. But we will know the reason of the test
failure.

BUG=chromium:1132444
TEST=tast run DUT cryptohome.GuestSession prints
the following messages:
'User session mount namespace has not been created.'
'Mount namespace is not ready.'

Cq-Depend:chromium:2596385
Change-Id: I339609c939efd09b29bb3b02c4508ff7b0d4b8f6