wifi: use RoamState for roam tests
$BUG prevented these tests from polling for the state of the Service, so
we relied on a BSS change signal from wpa_supplicant as a proxy. This
doesn't actually work because the BSS changes when we initiate an
association attempt, whereas we'd like to know when the association has
successfully completed. With the introduction of the RoamState property,
this is now easy to do. Switch over roam tests to use this property.

BUG=b:171086223
TEST=run affected tests

Change-Id: I9f7526e3a3250b3c06227611ae1fa918623940a1
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/2597788
Tested-by: Matthew Wang <matthewmwang@chromium.org>
Commit-Queue: Matthew Wang <matthewmwang@chromium.org>
Reviewed-by: Brian Norris <briannorris@chromium.org>
Reviewed-by: Yen-lin Lai <yenlinlai@google.com>