(Nearby Share) Add test for CB->Phone file transfer
Automated test for CB->phone Nearby Share file transfer using the API
approach (no UI interaction). The CrOS side is controlled by running
Javascript on a conn to chrome://nearby. All of the necessary properties
and methods are exposed to JS through the nearby-discovery-page and
nearby-confirmation-page elements.
The Android side is controlled via snippet RPC as described in
go/nearby-e2e-android-snippet. The RPCs required for the receiving flow
have been added here. The test has been implemented as a parameterized
test so we can extend it with different file types/combinations in the
future, the test data files simply need to be zipped and uploaded to the
private test assets gsbucket.

BUG=chromium:1129232
TEST=tast run <dut> nearbyshare.CrosSenderAndroidReceiver.small_png
TEST=tast run -var=rooted=false <dut>
nearbyshare.CrosSenderAndroidReceiver.small_jpg (passes on local unrooted
device)

Change-Id: I9daab09334af425ff57111b157298eeaaee8aedf