Add WebAuthn end-to-end test using external website
This test exercise the WebAuthn feature, which involves implementation
in Chrome, cryptohome, Ash, u2fd and cr50. The WebAuthn flow starts
from the browser receiving a WebAuthn request. The browser then talks
to u2fd to perform MakeCredential or GetAssertion.

Why this is not using a local js file: WebAuthn has security checks that
disallow using an IP address as host, or other "untrustworthy" domains.
Thus only an external website can invoke WebAuthn. This test uses
webauthn.io, which is a standard test website and is unlikely to change
very often.

BUG=b:169976548
TEST=tast -verbose run <IP> u2fd.WebauthnUsingPIN

Change-Id: Ic41457ac9a144f51188a5b0f3003a2d2561557e9