arc.PowerVideo*Perf: improve waiting on setup
Wait on the codec to be ready before starting measurements. This is
primarily for the encoder, as it has to read a very large file into
memory. This change also starts the metrics after the warmup sleep, as
duration based metrics tend to use Start() instead of StartRecording()
for their initial metric.

BUG=b:175515099
TEST=arc.PowerVideoDecodePerf.h264_1080p_30fps_vm
TEST=arc.PowerVideoEncodePerf.h264_1080p_i420_vm

Change-Id: Ia4ee9ccf9832df5e4d603970e7c92bfff1436f1f