health: add "chrome" SoftwareDep and restart Chrome in fixture
A running Chrome browser is required to maintain a mojo connection
between cros_healthd and Chrome. This change adds explicit dependencies
on Chrome for all cros_healthd tests. This includes the "chrome"
SoftwareDep and restarting the "ui" upstart service as a part of the
fixture and ensuring that cros_healthd properly bootstrapped.

BUG=b:178274447
TEST=(dut) stop ui
tast run sarien-dut health.* (pass)

Change-Id: Ibcb0006cc742590890e1596b7d5dedf43593a27a