wpasupplicant: Use ConnectNoTiming to avoid flooding timing.json.
Functions like BSSs will iterate all BSS object path. In open air
environment, it is possible this can create a lot of timing entries in
timing.json. Apply the similar pattern as in shill package to use
ConnectNoTiming in New and add timing on the functions that call
multiple New.

BUG=b:154618948
TEST=`tast run $DUT wifi.PMKSACaching` and check the "tf.DiscoverBSSID"
entry in timing.json

Change-Id: I5d893b68ecce116208ba5e2af299a8163eaaf1dc