Add PNaCl module execution test
Add a Tast test that verifies the basic functionality of PNaCl (Portable
Native Client).

The test loads the PNaCl binary module in the context of a Chrome App
(since [P]NaCl has been put behind an Origin Trial on the open web). The
test then verifies that a message can be sent to/from the module.

Despite that the [P]NaCl technology has been deprecated for a long time,
it's still used for some crucial applications and will stay so for a
while (until the final deprecation deadline comes), therefore the new
test should help preventing regressions in this area. This is becoming
more important now that the NaCl's overall test coverage decreases due
applications being slowly migrated away from it.

BUG=chromium:1162336
TEST=tast run $DUT nacl.Pnacl

Change-Id: I63a6762fa780d30b8f1a21bf220bd0efc1085883