cuj: add tab switch CUJ test cases
add tab_switch_cuj2.go. This test supports the CUJ running with
remote WPR Server. It is assumed that an external WPR server needs
to be set up with the WPR address provided to the tests via runtime
variables.

Also, a TabSwitchCUJRecorder2 test is provided to record web content
with local or remote WPR server

BUG=chromium:1196849
TEST=fast_build.sh -T
TEST=tast run -build=false -var=ui.cuj_username=$ACCOUNT -var=ui.cuj_password=$PASSWORD -var=wpr_http_addr=$WPR_ADDR -var=wpr_https_addr=$WPR_ADDR ui.TabSwitchCUJ2.*

Change-Id: Id34c226e2d3f89ef1cf6dc83377bf187cbf9f1e4