Split power-related data collector from cuj.Recorder
The error reports that a single 'snapshot' of timeline data can
take more than 1 second (the snapshot interval), which is treated
as an error.

One of the potential problem is actually addition of power metrics.
I suggested to simply merge them into cuj.Recorder, but this could
be a bad decision.  It is better to be split and recorded
separately.

Now they are recorded in a different timeline with 'Power' prefix,
so they won't interfere with each other.

BUG=chromium:1162174
TEST=tast run ui.MeetCUJ.4p ui.MeetCUJ.power_4p ui.TabSwitchCUJ

Change-Id: I239dcbb85a250ec7a1224b09528a1aee74bb4f4b