tast-tests: try multiple categories for camera labels
In HWID we are making changes to replace "video" category with "camera",
so the camera tast-test needs to be compatible with both names.  Note
that we only expect one of "video" and "camera" will appear in
host_info_labels but not both.

BUG=b:176400434
TEST=tast run -varsfile=labels.yaml $DUT platform.CrosRuntimeProbeCamera
with hwid_component:camera/<comp_name> labels.

Cq-Depend: chromium:2610671
Change-Id: I9e2675c0b0f95c97581433ef90baec5fa7460070