A speculative fix for ui.DesktopControl
The failure on juniper indicates that the quick settings tray
does not exist in the ui-tree, although it can be seen in the
screenshot. Putting a longer timeout for finding the collapse
button would increase the stability.

The failure of akali said that the TransitionToCollapsed exists
but TransitionToExtended does not exist, which means that
the extending transition does not finish in the scenario.
One possibility is that 'collapseButton.WaitLocationStable'
may finish earlier.  Increasing the 'interval' of checking
for WaitLocationStable will mitigate this issue, I believe.

BUG=chromium:1163981
TEST=tast run ui.DesktopControl on juniper and akali

Change-Id: Id8e34e2cdffdc03eb0e4e55e4c2cd6945b539aa5
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/2615775
Reviewed-by: Xiyuan Xia <xiyuan@chromium.org>
Tested-by: Jun Mukai <mukai@chromium.org>
Commit-Queue: Jun Mukai <mukai@chromium.org>