shill: Add shill_manager_sanity and shill_cellular_sanity tests
Adds sanity tests for the Shill Manager API and for Cellular.

This includes some additional Manager cleanup and functionality.
In particular, this changes Manager.FindMatchingService to no
longer return an err if no matching service is found, instead
nil,nil is return in that case. This behaves better with Poll
when waiting for a particular Service to appear.

BUG=b:177103167
TEST=tast --verbose run ${DUT} network.Shill*Sanity

Change-Id: Ie32783bbc242d5cf715d60d1422ac7d1c1986622